refactor: Move package of enum class
