--- src/main/java/http/util/constant/HttpHeader.java
@@ -1,15 +1,0 @@
-package http.util.constant;
-
-public enum HttpHeader {
-    CONTENT_TYPE("Content-Type"), CONTENT_LENGTH("Content-Length"), SET_COOKIE("Set-Cookie"), COOKIE("Cookie"), LOCATION("Location");
-
-    private String header;
-
-    HttpHeader(String header){
-        this.header = header;
-    }
-
-    public String getHeader() {
-        return header;
-    }
-}
