--- src/main/java/http/util/constant/HttpMethod.java
@@ -1,15 +1,0 @@
-package http.util.constant;
-
-public enum HttpMethod {
-    GET("GET"), POST("POST");
-
-    private String method;
-
-    HttpMethod(String method){
-        this.method = method;
-    }
-
-    public String getMethod() {
-        return method;
-    }
-}
